+++ owldata/__version__.py
@@ -7,4 +7,4 @@
 
 # =====================================================================
 
-__version__ = '0.0.14'
+__version__ = '0.0.15'
