+++ owldata/_owlerror.py
@@ -25,7 +25,8 @@ class OwlError():
                 'ColumnsError':'請重新確認您的欄位名稱是否輸入正確',
                 'ValueError':'日期規則錯誤，請重新確認!!',
                 'PdError':'無商品使用權限，請洽業務人員',
-                'ExError':'操作錯誤，請重新再輸入'
+                'ExError':'操作錯誤，請重新再輸入',
+                'CannotFind':'指定日期不在範圍內'
                 }
     
     _http_error = {
