+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
 name = owldata
-version = 0.0.14
+version = 0.0.15
 keywords = cmoney, financial data, owldata
 license = MIT
 desciption-file = README.md
