+++ setup.py
@@ -2,14 +2,12 @@ from setuptools import setup
 from owldata.__version__ import __version__
 import io
 
-
-
 with open('requirements.txt', encoding='utf-8') as fid:
     requires = [line.strip() for line in fid]
     
 # with open("README.md", encoding='utf-8') as f:
 #     long_description = f.read()
-    
+
 def read(path):
     with io.open(path, mode='r', encoding='utf-8') as f:
         return f.read()
